+++ tests/test_annotation.py
@@ -57,14 +57,6 @@ def adata_rna():
 
 # ------------------------- Tests ------------------------- #
 
-def test_add_cellxgene_annotation(adata_rna):
-    """Test if 'cellxgene' column is added to adata.obs."""
-
-    csv_f = os.path.join(os.path.dirname(__file__), 'data', 'cellxgene_anno.csv')
-    anno.add_cellxgene_annotation(adata_rna, csv_f)
-
-    assert "cellxgene_clusters" in adata_rna.obs.columns
-
 
 @pytest.mark.parametrize("inplace, threads, config, best, coordinate_cols",
                          [(True, 1, None, True, None),