+++ tests/tools/test_celltype_annotation.py
@@ -4,14 +4,28 @@ import os
 import pytest
 import anndata as ad
 from sctoolbox import celltype_annotation
+import sctoolbox.tools as anno
+import scanpy as sc
+
+# --------------------------- Fixtures ------------------------------ #
 
 
 @pytest.fixture
 def test_adata():
     """Load adata."""
-    adata_dir = os.path.join(os.path.dirname(__file__), 'data', 'scsa')
+    adata_dir = os.path.join(os.path.dirname(__file__), '..', 'data', 'scsa')
     adata = ad.read_h5ad(adata_dir + '/adata_scsa.h5ad')
     return adata
+
+
+@pytest.fixture
+def adata_rna():
+    """Load rna anndata."""
+    adata_f = os.path.join(os.path.dirname(__file__), '..', 'data', 'adata.h5ad')
+    return sc.read_h5ad(adata_f)
+
+
+# --------------------------- Tests --------------------------------- #
 
 
 def fetch_adata_uns(test_adata):
@@ -34,3 +48,12 @@ def test_run_scsa(test_adata, column):
     """Test run_scsa success."""
     adata = celltype_annotation.run_scsa(test_adata, species='Mouse', inplace=False, column_added=column)
     assert column in adata.obs.columns
+
+
+def test_add_cellxgene_annotation(adata_rna):
+    """Test if 'cellxgene' column is added to adata.obs."""
+
+    csv_f = os.path.join(os.path.dirname(__file__), '..', 'data', 'cellxgene_anno.csv')
+    anno.add_cellxgene_annotation(adata_rna, csv_f)
+
+    assert "cellxgene_clusters" in adata_rna.obs.columns